+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -4,6 +4,8 @@ WIP 5.13.2
 
 - Typings of `R.assoc` match its `@types/ramda` counterpart.
 
+- Simplify `R.forEach` typings
+
 5.13.1
 
 - Fix wrong `R.takeWhile`